Add krabs ETW headers & reorganize sources
Integrate the krabs ETW wrapper into the project and reorganize source/headers:

- Added a large set of krabs headers (krabs.hpp and many krabs/* headers) to include/krabs.
- Added etw_controller.cpp/.h and a new ulti/lru_cache.hpp header.
- Moved file_helper.* into include/ulti/ (renamed references accordingly).
- Removed legacy communication files: include/com/driver_comm.* and include/com/minifilter_comm.*.
- Updated RansomDetectorService.vcxproj and its .filters to reflect added/removed files and added <PreprocessToFile>false</PreprocessToFile>.

These changes wire the ETW-based krabs library into the build, clean up old COM/minifilter stubs, and reorganize utility code for reuse.

+++ RansomDetectorService/RansomDetectorService/include/manager/scanner.cpp
@@ -1,7 +1,7 @@
 ﻿#include "scanner.h"
 #include "receiver.h"
 #include "file_type_iden.h"
-#include "file_helper.h"
+#include "ulti/file_helper.h"
 #include "receiver.h"
 
 namespace manager {
@@ -129,9 +129,9 @@ namespace manager {
                 local_queue.push(std::move(tmp_queue.front()));
                 tmp_queue.pop();
             }
-            int n = local_queue.size();
+            size_t n = local_queue.size();
             // Process events
-            for (int i = 0; i < n; i++)
+            for (size_t i = 0; i < n; i++)
             {
                 if (local_queue.size() == 0) {
                     break;
